integrate PR feedback for #5397
