+++ frontend/javascripts/oxalis/model/actions/annotation_actions.js
@@ -6,7 +6,7 @@ import type {
   RemoteMeshMetaData,
   APIAnnotationVisibility,
 } from "types/api_flow_types";
-import type { Vector3 } from "oxalis/constants";
+import type { Vector3, AnnotationTool } from "oxalis/constants";
 import type { UserBoundingBox } from "oxalis/store";
 
 type InitializeAnnotationAction = {
@@ -43,6 +43,10 @@ type AddUserBoundingBoxesAction = {
   type: "ADD_USER_BOUNDING_BOXES",
   userBoundingBoxes: Array<UserBoundingBox>,
 };
+
+type SetToolAction = { type: "SET_TOOL", tool: AnnotationTool };
+
+type CycleToolAction = { type: "CYCLE_TOOL" };
 
 export type UpdateRemoteMeshMetaDataAction = {
   type: "UPDATE_REMOTE_MESH_METADATA",
@@ -123,6 +127,8 @@ export type AnnotationActionTypes =
   | SetAnnotationVisibilityAction
   | SetAnnotationDescriptionAction
   | SetAnnotationAllowUpdateAction
+  | SetToolAction
+  | CycleToolAction
   | UpdateRemoteMeshMetaDataAction
   | SetUserBoundingBoxesAction
   | AddUserBoundingBoxesAction
@@ -188,6 +194,15 @@ export const addUserBoundingBoxesAction = (
 ): AddUserBoundingBoxesAction => ({
   type: "ADD_USER_BOUNDING_BOXES",
   userBoundingBoxes,
+});
+
+export const setToolAction = (tool: AnnotationTool): SetToolAction => ({
+  type: "SET_TOOL",
+  tool,
+});
+
+export const cycleToolAction = (): CycleToolAction => ({
+  type: "CYCLE_TOOL",
 });
 
 export const updateRemoteMeshMetaDataAction = (
